+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added duploctl command to create AI Helpdesk Ticket & Send AI Message.
+
 ### Fixed
 
 - Now waiting for a service to update works with HPA enabled